pyproject.toml
[project]
name = "pdftotext-mnbvc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jsonlines>=4.0.0",
    "lingua-language-detector>=2.1.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "tqdm>=4.67.1",
]
